+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix nameless `subject` method completion inside nested `context` blocks
+
 ## [0.2.0] - 2024-05-20
 
 ### Added